+++ src/App.tsx
@@ -38,7 +38,7 @@ import {Modality, ModalityService} from "kk-modality";
 
 import './App.scss';
 
-export const C_VERSION = '0.36.5-call';
+export const C_VERSION = '0.36.6-call';
 export const C_ELECTRON_VERSIONS = ['10.1.1', '8.5.2'];
 
 export const isProd = (!process || !process.env || process.env.NODE_ENV !== 'development');